web - extracted match quick info scoreline into reusable component
