+++ web/src/components/features/match/match-quick-info/match-quick-info-card-content/match-quick-info-card-content.tsx
@@ -2,7 +2,7 @@ import { Separator } from "@/components/ui/separator";
 import { MatchQuickInfoEventsComponent } from "../match-quick-info-events/match-quick-info-events";
 import { IMatch } from "@/components/common/api-football-response";
 import { DateToolbox } from "@/components/common/toolbox/date";
-import { MatchToolbox } from "@/components/common/toolbox/match";
+import { MatchScorelineComponent } from "../../match-scoreline/match-scoreline";
 
 interface IMatchQuickInfoCardContentComponentProps {
     match: IMatch;
@@ -19,7 +19,7 @@ export const MatchQuickInfoCardContentComponent = (props: IMatchQuickInfoCardCon
                 </section>
 
                 <section className="flex items-center gap-2">
-                    <MatchQuickInfoCenterContent match={props.match} hasStarted={props.hasStarted} />
+                    <MatchQuickInfoCardCenterContentComponent match={props.match} hasStarted={props.hasStarted} />
                 </section>
 
                 <section className="flex-1 flex items-center justify-end gap-2">
@@ -33,36 +33,15 @@ export const MatchQuickInfoCardContentComponent = (props: IMatchQuickInfoCardCon
     );
 };
 
-const MatchQuickInfoCenterContent = (props: IMatchQuickInfoCardContentComponentProps) => {
+interface IMatchQuickInfoCardCenterContentComponent {
+    match: IMatch;
+    hasStarted: boolean;
+}
+
+const MatchQuickInfoCardCenterContentComponent = (props: IMatchQuickInfoCardCenterContentComponent) => {
     // display match scores if match has started
     if (props.hasStarted) {
-        const homeTeamScore: number = !!props.match.goals.home ? props.match.goals.home : 0;
-        const awayTeamScore: number = !!props.match.goals.away ? props.match.goals.away : 0;
-        const displayMatchStatus: string = MatchToolbox.getDisplayMatchStatus(props.match.fixture);
-        const matchInProgress: boolean = MatchToolbox.isMatchInProgress(props.match.fixture.status.short);
-
-        return (
-            <section className="flex flex-col items-center gap-2">
-                <section className="flex items-center gap-2 text-3xl font-bold">
-                    <div className={`${!matchInProgress && props.match.teams.away.winner && "opacity-40"}`}>
-                        {homeTeamScore}
-                        {props.match.fixture.status.short === "PEN" && (
-                            <span className="ml-1">({props.match.score.penalty.home})</span>
-                        )}
-                    </div>
-                    <p className="opacity-40">-</p>
-                    <div className={`${!matchInProgress && props.match.teams.home.winner && "opacity-40"}`}>
-                        {awayTeamScore}
-                        {props.match.fixture.status.short === "PEN" && (
-                            <span className="ml-1">({props.match.score.penalty.away})</span>
-                        )}
-                    </div>
-                </section>
-                <section className="font-bold">
-                    <p>{displayMatchStatus}</p>
-                </section>
-            </section>
-        );
+        return <MatchScorelineComponent match={props.match} />;
     }
 
     const matchDate = new Date(props.match.fixture.date);
